Nodify _config usage medic/medic-webapp#2855
We're just going to presume one node here, because they are migrations
that are run once, and are solving legacy problems that will only have
existed on that first local node.

If we ever start running a setup with multiple nodes, none of which are
on the same box as medic-api, we may have to revisit this.