Using sinon.sandbox for clearer, safer tests
Replacing our pattern of passing a list of things that have maybe been
stubbed to a utility that restores them, and replacing it with sinon
sandboxes, the correct way to deal with this kind of thing.

This is better because it's fool-proof, you do not have to know what
things may or may not have been stubbed.

See: http://sinonjs.org/releases/v2.1.0/sandbox/sinon