Split couch _changes requests with _doc_ids filter into batches of 100
Using the `_doc_ids` filter with a `_changes` request with over 100
doc ids runs a non-optimised path in CouchDB.
  See: https://issues.apache.org/jira/browse/COUCHDB-1288

100 or less documents uses a much faster code path.

To make sure we use the fast path as much as possible, we create N
`_changes` requests, in batches of 100 docs, and combine them before
passing it back to the client.

Note: we are not worried about the documented reasons for the fast path
being limited to 100 documents. Most of our documents are not heavily
edited.

Note: we don't do this for longpoll requests, because of the complexity
/ ambiguity in managing multiple open longpoll requests, and the risk of
potentially skipping / missing documents by processing them in the wrong
order. Longpoll requests are by definition not as taxing on the system
anyway, as you are now at the point where you are waiting for changes to
occur for you to pick up, as opposed to pulling hundreds of documents
because you're setting up for the first time / restarting the app.

medic/medic-webapp#3508