Don't throw errors for bad messages medic/medic-webapp#3672
Instead of erroring simply log the problems and otherwise ignore the bad
data. We want to do this because we don't want to push problems back to
gateway, as gateway is hard to debug or even access, whereas api is in a
nice easy location to view logs and so on.