Add server whitelist support
Adds support for reading an array in /tmp/whitelist.json. If the file
exists it should be an array of usernames that are currently allowed to
use the service. If it doesn't exist that means everyone can use it.

/tmp/whitelist.json
```
["chw1", "chw2"]
```

Would allow usernames `chw1` and `chw2` to use the service, and no other
 users. However, admins are always allowed to use the service.

In this instance "using the service" means any authenticated interaction
through 5988 at all, which also includes any external CouchDB request
attempts. In this version Sentinel access CouchDB directly (via 5984)
 and should not be affected (and is also using an admin user).